+++ src/main/java/com/ventionteams/applicationexchange/repository/LotRepository.java
@@ -4,6 +4,6 @@ import com.ventionteams.applicationexchange.entity.Lot;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface LotRepository extends JpaRepository<Lot, Integer> {
+public interface LotRepository extends JpaRepository<Lot, Long> {
     List<Lot> findAllByCategoryId(Integer id);
 }
